test_task_2/pyproject.toml
[tool.poetry]
name = "test-task-2"
version = "0.1.0"
description = ""
authors = ["Arsen"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = {extras = ["email"], version = "^2.7.0"}
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
